integrate docker-compose prepared for depoloy

+++ history_api/models.py
@@ -24,7 +24,7 @@ class Trades(models.Model):
     asker = models.CharField(max_length=12, null=True)
     bid = models.CharField(max_length=20, null=True)
     ask = models.CharField(max_length=20, null=True)
-    price = models.IntegerField(null=True, default=0)
+    price = models.BigIntegerField(null=True, default=0)
     quote = models.CharField(max_length=300, null=True)
     base = models.CharField(max_length=300, null=True)
     time = models.DateTimeField(null=True)
@@ -34,7 +34,7 @@ class Trades(models.Model):
 
 
 class PriceHistory(models.Model):
-    price = models.IntegerField(null=True, default=0)
+    price = models.BigIntegerField(null=True, default=0)
     quote = models.CharField(max_length=300, null=True)
     base = models.CharField(max_length=300, null=True)
     time = models.DateTimeField(null=True)
@@ -47,11 +47,9 @@ class PriceHistory(models.Model):
 
 
 class Market(models.Model):
-    market_id = models.IntegerField(default=0, null=True)
+    market_id = models.BigIntegerField(default=0, null=True)
     quote = models.CharField(max_length=12)
     base = models.CharField(max_length=12)
-    #quote = JSONField(max_length=500, default=default_asset)
-    #base = JSONField(max_length=500, default=default_asset)
     contract = models.CharField(max_length=12, null=True)
 
 
@@ -63,10 +61,10 @@ class SellOrder(models.Model):
     bid = models.DecimalField('Currency Amount', default=0, null=True, max_digits=19, decimal_places=10)
     ask = models.DecimalField('Currency Amount', default=0, null=True, max_digits=19, decimal_places=10)
 
-    price = models.IntegerField(default=0, null=True)
+    price = models.BigIntegerField(default=0, null=True)
     time = models.DateTimeField(null=True)
 
-    order_id = models.IntegerField(null=True)
+    order_id = models.BigIntegerField(null=True)
 
     def __str__(self):
         return f'(SellOrder) {self.account} ask: {self.ask} {self.base} | bid: {self.bid} {self.quote}'
@@ -80,9 +78,9 @@ class BuyOrder(models.Model):
     bid = models.DecimalField('Currency Amount', default=0, null=True, max_digits=19, decimal_places=10)
     ask = models.DecimalField('Currency Amount', default=0, null=True, max_digits=19, decimal_places=10)
 
-    price = models.IntegerField(default=0, null=True)
+    price = models.BigIntegerField(default=0, null=True)
     time = models.DateTimeField(null=True)
-    order_id = models.IntegerField(null=True)
+    order_id = models.BigIntegerField(null=True)
 
     def __str__(self):
         return f'(BuyOrder) {self.account} ask: {self.ask} {self.quote} | bid: {self.bid} {self.base}'
